[REFACTOR][IR] Initialize Unified IR Pass Infra.
Move the relay's pass Infra to ir.
Keep FunctionPass in relay as it is local to the dialect.
